Level 5
A finished Level 5 (unless I've forgotten something) minus the lighting.
I've added grass rock trees a fade out, a player reset if they don't bring the first cube with them.
A new background for the level select in the main menu so the back button wont be overlapping with the word Power.